added unit tests for Modal component
